🎨 Rename module: goframe -> goframe/app
This change sets the default executable name to "app".
We may also drop the `-o app` in the Dockerfile.
