+++ frameworks/Go/goframe/src/main.go
@@ -4,8 +4,9 @@ import (
 	"net/http"
 	"runtime"
 
+	"goframe/app/handler"
+
 	"github.com/gogf/gf/v2/frame/g"
-	"goframe/handler"
 )
 
 const (
